# target/classes/openapi.yaml
openapi: 3.0.3
info:
  title: DigiShopKart
  description: Apis of E-commerce application
  contact:
    name: Ajay Kate
    url: http://digishopkart.com
  version: '0.1'
servers:
  - url: http://localhost:8085
#------------------------------PATHS ---------------------------------
paths:
...
  /api/v1/variant:
    get:
      tags:
        - Variant
      summary: fetch all Variants
      operationId: fetchAllVariants
      description: API to fetch all Variants
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Variant'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/variant/{productId}:
    post:
      tags:
        - Variant
      summary: insert a Variant
      operationId: insertVariant
      description: API to insert Variant record
      requestBody:
        $ref: '#/components/requestBodies/VariantBody'
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Variant'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'
  /api/v1/variant/{variantId}:
    get:
      tags:
        - Variant
      summary: fetch a Variant
      operationId: fetchVariant
      description: API to fetch the Variant
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Variant'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Variant
      summary: update a Variant
      operationId: updateVariant
      description: API to update the Variant by variantId
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/VariantBody'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Variant'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Variant
      summary: delete a Variant
      operationId: deleteVariant
      description: API to delete Variant by variantId
      parameters:
        - in: path
          name: variantId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Variant'
        404:
          $ref: '#/components/responses/NotFound'
        401:
          $ref: '#/components/responses/Unauthorized'

  #  ********************************** DISCOUNT **********************************
...
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
...
  requestBodies:
...
    VariantBody:
      description: A JSON object Contenting Variant
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Variant'
...
  schemas:
    Errors:
      type: object
      properties:
        code:
          type: string
        message:
          title: string
      required:
        - code
        - message
...
    Variant:
      type: object
      required:
        - colour
        - size
      properties:
        id:
          type: integer
        name:
          type: string
          example: Size, Colour, etc...
        value:
          type: string
        status :
          type: string
          enum:
            - active
            - inactive
...